Determine aspect ratio based on node dimensions
