+++ templates/vue/src/components/TapestryApp.vue
@@ -15,7 +15,7 @@ import client from "../services/TapestryAPI"
 import { names } from "@/config/routes"
 import Toolbar from "./Toolbar"
 import TapestryMain from "./TapestryMain"
-import { mapMutations, mapState } from "vuex"
+import { mapGetters, mapMutations, mapState } from "vuex"
 import TapestryMap from "./TapestryMap"
 import Helpers from "@/utils/Helpers"
 
@@ -32,6 +32,7 @@ export default {
   },
   computed: {
     ...mapState(["nodes", "links", "selection", "settings", "rootId"]),
+    ...mapGetters(["getNodeDimensions"]),
     isSidebarOpen() {
       return Boolean(this.$route.query.sidebar)
     },
@@ -76,7 +77,7 @@ export default {
             width,
             height,
           } = this.$refs.graph.$refs.app.getBoundingClientRect()
-          const { x0, y0, x, y } = this.getNodeDimensions()
+          const { x0, y0, x, y } = this.getNodeDimensions
 
           const tapestryDimensions = {
             startX: 0,
@@ -120,25 +121,6 @@ export default {
           )}`
         }
       }
-    },
-    getNodeDimensions() {
-      const box = {
-        x0: 30000,
-        y0: 30000,
-        x: 0,
-        y: 0,
-      }
-      for (const node of Object.values(this.nodes)) {
-        if (node.nodeType !== "") {
-          const { x, y } = node.coordinates
-          box.x0 = Math.min(x, box.x0)
-          box.y0 = Math.min(y, box.y0)
-          box.x = Math.max(x, box.x)
-          box.y = Math.max(y, box.y)
-        }
-      }
-
-      return box
     },
     openNode(id) {
       this.$router.push({
