+++ templates/vue/src/store/getters.js
@@ -202,3 +202,23 @@ export function getNeighbouringLinks(state) {
 export function getTheme(state) {
   return state.theme ? state.theme : "light"
 }
+
+export function getNodeDimensions(state) {
+  const box = {
+    x0: 30000,
+    y0: 30000,
+    x: 0,
+    y: 0,
+  }
+  for (const node of Object.values(state.nodes)) {
+    if (node.nodeType !== "") {
+      const { x, y } = node.coordinates
+      box.x0 = Math.min(x, box.x0)
+      box.y0 = Math.min(y, box.y0)
+      box.x = Math.max(x, box.x)
+      box.y = Math.max(y, box.y)
+    }
+  }
+
+  return box
+}
